+++ localization.fr.lua
@@ -19,7 +19,7 @@ if (GetLocale() == "frFR") then
 	AUTOFLOOD_HELP = {
 		"===================== Auto Flood =====================",
 		"/flood [on|off] : Démarre / arrête l'envoi du message.",
-		"/floodmsg <message> : Définit le message à envoyer. Utilisez {size} pour afficher la taille actuelle du groupe/raid.",
+		"/floodmsg <message> : Définit le message à envoyer. Utilisez {size} pour la taille du groupe, {tanks} pour le nombre de tanks, {heals} pour le nombre de soigneurs, et {dps} pour le nombre de DPS.",
 		"/floodchan <canal> : Définit le canal à utiliser pour l'envoi.",
 		"/floodrate <durée> : Définit la période (en secondes) d'envoi du message.",
 		"/floodinfo : Affiche les paramètres.",